add Slide animation for ImageListItems
