+++ src/routes/home.tsx
@@ -1,6 +1,6 @@
 import { Container, Typography } from '@mui/material';
-import AppBar from '../components/appBar';
-import StandardImageList from '../components/standardImageList';
+import AppBar from '../components/AppBar';
+import StandardImageList from '../components/StandardImageList';
 
 function Home() {
   return (
@@ -13,6 +13,17 @@ Everyone is entitled in full equality to a fair and public hearing by an indepen
 No one shall be subjected to arbitrary interference with his privacy, family, home or correspondence, nor to attacks upon his honour and reputation. Everyone has the right to the protection of the law against such interference or attacks.
         </Typography>
         <StandardImageList />
+        <Typography 
+        sx={{
+          p: 5,
+          m: 3, 
+          textAlign: 'center',
+          fontSize:'48px', 
+          color: 'rgb(129, 129, 129)'
+        }}
+          >
+          Footer placeholder
+        </Typography>
       </Container>
     </>
   )
